backend/requirements.txt
Django==5.1.3
django-environ==0.11.2
typing_extensions==4.12.2
pytest==8.3.3
ruff==0.6.9
psycopg2-binary==2.9.9
django-types==0.19.1
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support


